requirements.txt
langchain
langchain-openai
langgraph
tavily-python
pydantic
python-dotenv
yfinance
pandas